+++ lib/common.h
@@ -7,9 +7,6 @@
 #include "prj.h"
 #include "limit.h"
 
-#define PROGRAM     "tman"
-#define VERSION     "v0.0.6"
-
 #define TRUE        1
 #define FALSE       0